Implemented Filled line plot for distributions. Removed Edit menu, since it wasn't implemented at all. Tweaked volume density scaling. Fixed single '*' for range which results in (-1,-1) interpreted as undef. to undef. Changed immediate mode index type to uint32 from uint16. Fixed bug in stats when computing undefined range for atoms.

+++ src/gfx/immediate_draw_utils.cpp
@@ -14,7 +14,7 @@ struct Vertex {
     uint32 color;
 };
 
-using Index = uint16;
+using Index = uint32;
 
 struct DrawCommand {
     Index offset;
